requirements-dev.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1
django-cors-headers==3.5.0
django-filter==2.3.0
django-heroku==0.3.1
djangorestframework==3.11.1
importlib-metadata==2.0.0
Markdown==3.2.2
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
static3==0.7.0
validate-docbr==1.7.0
whitenoise==5.2.0
zipp==3.1.0
